% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pyscvi_version}
\alias{check_pyscvi_version}
\title{Check that the current scvi-tools version in Python is up to date.}
\usage{
check_pyscvi_version()
}
\description{
Check that the current scvi-tools version in Python is up to date.
}
