% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scvi_model.R
\name{CreateSCVI}
\alias{CreateSCVI}
\alias{CreateSCVI.Seurat}
\title{Create and train SCVI model}
\usage{
CreateSCVI(object, ...)

\method{CreateSCVI}{Seurat}(
  object,
  main_layer = "counts",
  max_epochs = NULL,
  use_gpu = NULL,
  train = TRUE,
  early_stopping = FALSE
)
}
\arguments{
\item{max_epochs}{Number of passes through the dataset.
If NULL, defaults to min(round(c(round((20000 / n_cells) * 400), 400)))}

\item{use_gpu}{Use default GPU if available (if NULL or TRUE),
or index of GPU to use (if int),
or name of GPU (if string), or use CPU (if FALSE).}

\item{train}{If TRUE the model will be created and trained.}

\item{early_stopping}{Perform early stopping.}
}
\value{
Return a trained SCVI model.
}
\description{
This function create and train a scvi model.
}
\concept{Create scvi-tools model}
