% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scvi_model.R
\name{RunSCVI}
\alias{RunSCVI}
\alias{RunSCVI.Seurat}
\title{scvi-tools single cell}
\usage{
RunSCVI(object, ...)

\method{RunSCVI}{Seurat}(
  object,
  scvi_model = NULL,
  reduction.name = "scvi",
  reduction_key = "scvi_"
)
}
\arguments{
\item{object}{A seurat object}

\item{...}{other parameters}

\item{scvi_model}{A scvi_model}
}
\value{
Return a Seurat object with the computed embedding
        with all the measure computed. For downstream Seurat analyses,
        use reduction='scvi'.
}
\description{
Run dimentionality reduction usinc SCVI model
}
\concept{Run scvi-tools model}
